add linting script and some vscode settings

+++ src/rules/IRegistryRule.sol
@@ -1,7 +1,12 @@
-// SPDX-License-Identifier: GPL-3 
+// SPDX-License-Identifier: GPL-3
 pragma solidity ^0.8.4;
 
 interface IRegistryRule {
     function canRegister(string calldata _label, address _sender) external returns (bool);
-    function canRegisterWithToken(string calldata _label, address _sender, uint _tokenId) external returns (bool);
+
+    function canRegisterWithToken(
+        string calldata _label,
+        address _sender,
+        uint256 _tokenId
+    ) external returns (bool);
 }
